+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Runnable reference applications for RAGOps."""